chore: 스토리북 에러 해결 및 폴더 구조 정리
