+++ app/(main)/store/home/_components/register-modal/RegisterModal.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import RegisterModal from './RegisterModal';
 
 const meta: Meta<typeof RegisterModal> = {
-  title: 'Components/RegisterModal',
+  title: 'Modals/RegisterModal',
   component: RegisterModal,
   tags: ['autodocs'],
 };
